project11/website/backend/requirements.txt
fastapi==0.104.1
hypercorn==0.14.4
pydantic
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.1

